Branch

+++ lms/deleteschool.php
@@ -1,20 +1,26 @@
 <?php
-        require('connect.php');
-        
-        if(isset($_POST['id'])) {
-            // print_r($POST);
-            $id = $_POST['id'];
+    session_start();
 
-            $query = "DELETE FROM schools WHERE id=$id";
-
-        $school = mysqli_query($connect, $query);
-        
-        if($school) {
-            echo "Record deleted successfully";
-            header('Location: index.php');
-        }
-        else {
-            echo 'FAILED, error code is' . mysqli_error($connect);
-        }
+    if (!isset($_SESSION['user_id']))
+    {
+        header("Location: login.php");
+        exit();
     }
-    ?>
+    require('connect.php');
+    
+    if(isset($_POST['id'])) {
+        // print_r($POST);
+        $id = $_POST['id'];
+
+        $query = "DELETE FROM schools WHERE id=$id";
+
+    $school = mysqli_query($connect, $query);
+    
+    if($school) {
+        header('Location: index.php');
+    }
+    else {
+        echo 'FAILED, error code is' . mysqli_error($connect);
+    }
+}
+?>
